[IMP] mail: move fileUploader methods in the model view
task-2579306
