+++ addons/mail/static/src/components/file_uploader/file_uploader.xml
@@ -3,7 +3,9 @@
 
     <t t-name="mail.FileUploader" owl="1">
         <div class="o_FileUploader" t-attf-class="{{ className }}" t-ref="root">
-            <input class="o_FileUploader_input" t-on-change="_onChangeAttachment" multiple="true" type="file" t-ref="fileInput" t-key="'fileInput'"/>
+            <t t-if="fileUploaderView">
+                <input class="o_FileUploader_input" t-on-change="fileUploaderView.onChangeAttachment" multiple="true" type="file" t-ref="fileInput" t-key="'fileInput'"/>
+            </t>
         </div>
     </t>
