+++ addons/mail/static/src/models/activity_view/activity_view.js
@@ -61,7 +61,7 @@ registerModel({
          * explorer for upload.
          */
         onClickUploadDocument() {
-            this.fileUploaderView.component.openBrowserFileUploader();
+            this.fileUploaderView.openBrowserFileUploader();
         },
         /**
          * @private
